Decouple GlowingDrawable from GlowingSpriteText
